# tests/openapi.yaml
openapi: 3.0.3
info:
  title: Tiikii Festival API
  description: API para el sistema de festivales Tiikii con funcionalidades de tickets, chat, mesh network y más
  version: 1.0.0
  contact:
    name: Tiikii Development Team
    email: dev@tiikii.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api.tiikii.com
    description: Servidor de producción
security:
  - BearerAuth: []
paths:
...
  /api/chat/rooms:
    get:
      summary: Obtener salas de chat del usuario
      description: Lista todas las salas de chat del usuario autenticado
      responses:
        '200':
          description: Lista de salas de chat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatRoom'

    post:
      summary: Crear sala de chat privada
      description: Crea una nueva sala de chat privada con otro usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId:
                  type: string
                  example: "user-456"
                festivalId:
                  type: string
                  example: "festival-123"
      responses:
        '201':
          description: Sala de chat creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Chat room created successfully"
                  chatRoom:
                    $ref: '#/components/schemas/ChatRoom'
  /api/chat/rooms/{roomId}/messages:
    get:
      summary: Obtener mensajes de sala
      description: Obtiene todos los mensajes de una sala de chat
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
          example: "room-123"
      responses:
        '200':
          description: Lista de mensajes
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId:
                    type: string
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'

    post:
      summary: Enviar mensaje
      description: Envía un mensaje a una sala de chat
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
          example: "room-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: "Hello everyone!"
      responses:
        '201':
          description: Mensaje enviado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message sent successfully"
                  chatMessage:
                    $ref: '#/components/schemas/ChatMessage'
  /api/chat/rooms/{roomId}/read:
    put:
      summary: Marcar mensajes como leídos
      description: Marca todos los mensajes de una sala como leídos
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
          example: "room-123"
      responses:
        '200':
          description: Mensajes marcados como leídos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Messages marked as read"
  /api/chat/messages/{messageId}:
    put:
      summary: Editar mensaje
      description: Edita un mensaje existente
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          example: "message-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: "Updated message text"
      responses:
        '200':
          description: Mensaje actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message updated successfully"

    delete:
      summary: Eliminar mensaje
      description: Elimina un mensaje (eliminación suave)
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          example: "message-123"
      responses:
        '200':
          description: Mensaje eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message deleted successfully"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatRoom:
      type: object
      properties:
        id:
          type: string
          example: "room-123"
        festivalId:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "General Chat"
        type:
          type: string
          enum: [general, friends, group]
          example: "general"
        avatar:
          type: string
          example: "👥"
        createdBy:
          type: string
          example: "user-123"
        unreadCount:
          type: integer
          example: 5
        lastMessage:
          type: string
          example: "Hello everyone!"
        lastMessageTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          example: "message-123"
        roomId:
          type: string
          example: "room-123"
        senderId:
          type: string
          example: "user-123"
        text:
          type: string
          example: "Hello everyone!"
        timestamp:
          type: string
          format: date-time
        isEdited:
          type: boolean
          example: false
        editedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
          example: false
        deletedAt:
          type: string
          format: date-time
          nullable: true
        senderName:
          type: string
          example: "John Doe"
        isOwn:
          type: boolean
          example: true
...